remove styles.js
